refactor(chalice): use CH to get replay events
